+++ plugins/modules/na_ontap_vscan.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp Inc.
+# (c) 2018-2026, NetApp Inc.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 '''
@@ -22,8 +22,6 @@ extends_documentation_fragment:
     - netapp.ontap.netapp.na_ontap
 version_added: 2.9.0
 author: NetApp Ansible Team (@carchi8py) <ng-ansible-team@netapp.com>
-notes:
-- on demand task, on_access_policy and scanner_pools must be set up before running this module
 description:
 - Enable and Disable Vscan
 options:
@@ -38,6 +36,32 @@ options:
     - the name of the data vserver to use.
     required: true
     type: str
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - On demand task, on_access_policy and scanner_pools must be set up before running this module.
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -82,9 +106,11 @@ class NetAppOntapVscan(object):
             enable=dict(type='bool', default=True),
             vserver=dict(required=True, type='str'),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[['use_lambda', True, ('lambda_config',)]],
         )
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
@@ -94,6 +120,8 @@ class NetAppOntapVscan(object):
         if self.rest_api.is_rest():
             self.use_rest = True
         else:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if HAS_NETAPP_LIB is False:
                 self.module.fail_json(msg="the python NetApp-Lib module is required")
             else: