+++ plugins/modules/na_ontap_vscan_on_access_policy.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp Inc.
+# (c) 2018-2026, NetApp Inc.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -101,6 +101,31 @@ options:
       - If not specified, default value is False in creating policy.
     type: bool
     version_added: 21.20.0
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -177,13 +202,15 @@ class NetAppOntapVscanOnAccessPolicy:
             scan_readonly_volumes=dict(required=False, type='bool'),
             only_execute_access=dict(required=False, type='bool')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             mutually_exclusive=[
                 ['filters', 'scan_readonly_volumes'],
                 ['filters', 'only_execute_access']
-            ]
+            ],
+            required_if=[['use_lambda', True, ('lambda_config',)]],
         )
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
@@ -213,6 +240,8 @@ class NetAppOntapVscanOnAccessPolicy:
                     self.parameters['scan_readonly_volumes'] = True
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])