+++ plugins/modules/na_ontap_vscan_on_demand_task.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -122,6 +122,31 @@ options:
     - Name of the task.
     type: str
     required: True
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -189,11 +214,13 @@ class NetAppOntapVscanOnDemandTask:
             schedule=dict(required=False, type="str"),
             task_name=dict(required=True, type="str")
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             required_if=[
-                ["state", "present", ["report_directory", "scan_paths"]]
+                ["state", "present", ["report_directory", "scan_paths"]],
+                ['use_lambda', True, ('lambda_config',)],
             ]
         )
         self.na_helper = NetAppModule()
@@ -204,6 +231,8 @@ class NetAppOntapVscanOnDemandTask:
                                        'scan_priority']
         self.use_rest = self.rest_api.is_rest_supported_properties(self.parameters, unsupported_rest_properties)
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if self.parameters.get('cross_junction') is None:
                 self.parameters['cross_junction'] = False
             if self.parameters.get('directory_recursion') is None: